Repair showcase compilation and close its rendering holes
The merge into master left the tree uncompilable and the showcase full of
permanent geometry holes. This gets both back.

Compilation (158 errors across 12 assemblies -> 0 across 72):

- 3a91a8f8 clobbered DecorationModel.cs, replacing the 576-line model with a
  316-line draft while claiming to add one enum value. That dropped
  DecorationSeed, DecorationExclusionKind, GeneratedPropIds, DecorationValidation,
  DecorationPropPresets and BedroomSceneDefinition, which most of
  Game.Structures.Runtime consumes. Restored the model and re-applied the
  ExteriorYard addition on top, then reverted 6c77ad97, which had rewritten
  WorldObjectIds against the truncated 128-bit GeneratedPropId.
- Game.Structures.Runtime and .Tests were missing their Unity.Collections
  reference; the config structs expose FixedList512Bytes through their API.
- Unity ships NUnit 3.5, which predates Assert.Multiple (3.6). ~50 test files
  call it, so each of the two affected assemblies gets a shim that shadows
  NUnit.Framework.Assert and inherits everything else.
- Assorted drift: SecondaryMaterial -> SoftMaterial, missing usings, the
  brick-pool widening that left four callers on int, and a CastlePlan wrapper
  that cannot convert through an `in` parameter.

Showcase bake:

Decoration placement gave up after four seeded offsets per socket, so a
required prop could fail in a heavily excluded room and abort the whole
bedchamber - which aborted the castle, which aborted the bake. Added a
deterministic sweep of the remaining offsets that runs only after every seeded
attempt has failed, so existing placements keep their exact positions.

Rendering holes (~1,300 visible chunks with no geometry, permanently):

- SurfaceGeometryArena was a fixed 2M verts / 6M indices (88 MB) holding ~141
  chunks, against ~1,480 in a production frustum. Per-chunk cost is not the
  defect: at 0.1 m voxels a 64-cell chunk spans 6.4 m and its faceted surface is
  ~12.5K vertices. The arena is now sized from DeviceTierBudget, split at the
  measured 1.51 indices per vertex rather than a fixed 3.0 that wasted half the
  index buffer while the vertex buffer starved.
- Arena pressure only ever evicted offscreen leases, so once the resident set
  was entirely in frustum it could never free one: publications failed to
  acquire, builds stopped, and the view froze with holes. Added TryEvictOneBehind,
  retiring the farthest published lease behind the nearest chunk waiting to
  publish, so the near field wins the arena.
- Per-frame upload budgets were far below what their own millisecond deadlines
  allow: 1 MiB against ~476 KB chunks admitted two chunks a frame.
- Startup bake covered 153.6 m while generation needs ~18 s to reach 200 m, so
  the player could walk into ground still being generated. Bake radius 3 -> 5.

Tests:

SilhouetteHoleRegressionTests asserted after a single warmup frame following a
camera teleport, so it could only ever observe an empty scheduler - it passed
vacuously or failed on visible=0. It now drives the view to convergence and
requires four consecutive hole-free frames. It converges at frame 1360 with
missing=0, visible=1239, resident=1539 and zero arena allocation failures.

FarTerrainHeightsMatchTheHeightField compared against the deliberate flat
base-height startup square. VoxelFarTerrain now exposes
HasSampledHeightsForEveryRing and the test waits for real heights.

+++ Assets/Game/Composition/Showcase/ShowcaseWorld.cs
@@ -273,7 +273,7 @@ namespace VoxelEngine.Showcase
 
         public ShowcaseWorld(uint seed, int brickPoolCapacity, int loadRadiusRegions,
                              int unloadRadiusRegions,
-                             int maxMixedBrickAllocationBytes =
+                             long maxMixedBrickAllocationBytes =
                                  VoxelEngineBootstrap.MaximumMixedBrickAllocationBytes)
         {
             Seed = seed;
